Extract styles.tsx and refactor some other things
